refactor(skills): extract publish dialogs
